requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
bidict==0.23.1
frozenlist==1.4.1
h11==0.14.0
idna==3.6
multidict==6.0.5
python-engineio==4.9.0
python-socketio==5.11.1
simple-websocket==1.0.0
wsproto==1.2.0
yarl==1.9.4